feat : 당첨번호를 입력하면 Contains를 활용해 당첨 번호를 체크한다.

+++ src/main/java/view/ResultView.java
@@ -3,14 +3,13 @@ package view;
 import domain.Lotto;
 
 import java.util.List;
-import java.util.Scanner;
 
 public class ResultView {
 
     private ResultView() {
     }
 
-    public static void printLottoCount(Long count) {
+    public static void printLottoCount(int count) {
         System.out.println(count + "개를 구매했습니다.");
     }
